Refactoring of search model
