+++ youtube-analyzer.py
@@ -1,14 +1,7 @@
 import sys
-import operator
 import csv
-import traceback
-from youtubesearchpython import (
-    VideosSearch, 
-    Channel
-)
 from PySide6.QtCore import (
-    Qt,
-    QAbstractTableModel
+    Qt
 )
 from PySide6.QtGui import (
     QAction,
@@ -33,81 +26,16 @@ from PySide6.QtWidgets import (
     QMessageBox
 )
 import xlsxwriter
+from model import (
+    ResultTableModel
+)
+from engine import (
+    YoutubeGrepEngine
+)
 
 
 app_name = "YouTube Analyzer"
-version = "1.1"
-
-
-def view_count_to_int(count_str):
-    if not count_str:
-        return 0
-    parts = count_str.split()
-    if len(parts) == 0:
-        return 0
-    processed_count = parts[0].replace(",", "")
-    return int(processed_count) if processed_count.isdigit() else 0
-
-
-def subcriber_count_to_int(count_str):
-    if not count_str:
-        return 0
-    parts = count_str.split()
-    if len(parts) == 0:
-        return 0
-    number_letter = parts[0]
-    match number_letter[-1]:
-        case "K":
-            multiplier = 1000
-        case "M":
-            multiplier = 1000000
-        case "B":
-            multiplier = 1000000000
-        case _:
-            multiplier = 1
-    return int(float(number_letter[:-1]) * multiplier)
-
-
-class ResultTableModel(QAbstractTableModel):
-    def __init__(self, parent, header, *args):
-        QAbstractTableModel.__init__(self, parent, *args)
-        self.result = []
-        self.header = header
-
-    def setData(self, result):
-        self.beginResetModel()
-        self.result = result
-        self.endResetModel()
-
-    def clear(self):
-        self.beginResetModel()
-        self.result.clear()
-        self.endResetModel()
-
-    def rowCount(self, parent):
-        return len(self.result)
-
-    def columnCount(self, parent):
-        return len(self.header)
-
-    def data(self, index, role):
-        if not index.isValid():
-            return None
-        elif role != Qt.ItemDataRole.DisplayRole:
-            return None
-        return self.result[index.row()][index.column()]
-
-    def headerData(self, col, orientation, role):
-        if orientation == Qt.Orientation.Horizontal and role == Qt.ItemDataRole.DisplayRole:
-            return self.header[col]
-        return None
-
-    def sort(self, col, order):
-        self.layoutAboutToBeChanged.emit()
-        self.result = sorted(self.result, key=operator.itemgetter(col))
-        if order == Qt.SortOrder.DescendingOrder:
-            self.result.reverse()
-        self.layoutChanged.emit()
+version = "2.0dev"
 
 
 class AboutDialog(QDialog):
@@ -196,12 +124,7 @@ class MainWindow(QMainWindow):
         h_layout.addWidget(self._search_button)
 
         v_layout = QVBoxLayout()
-        header = ["Title", "Published Time", "Duration", 
-                  "View Count", "Link", 
-                  "Channel Name", "Channel Link", "Channel Subscribers",
-                  "Channel Views", "Channel Joined Date",
-                  "Views/Subscribers"]
-        self._model = ResultTableModel(self, header)
+        self._model = ResultTableModel(self)
         self._table_view = QTableView(self)
         self._table_view.setModel(self._model)
         self._table_view.resizeColumnsToContents()
@@ -236,37 +159,15 @@ class MainWindow(QMainWindow):
         self._model.clear()
         QApplication.instance().processEvents()
 
-        try:
-            videos_search = VideosSearch(self._request_text, limit = request_limit)
-            result = []
-            has_next_page = True
-            counter = 0
-            while has_next_page:
-                result_array = videos_search.result()["result"]
-                for video in result_array:
-                    views = view_count_to_int(video["viewCount"]["text"])
-                    channel = Channel.get(video["channel"]["id"])
-                    channel_views = view_count_to_int(channel["views"])
-                    channel_subscribers = subcriber_count_to_int(channel["subscribers"]["simpleText"])
-                    result.append((video["title"], video["publishedTime"], video["duration"], 
-                                views, video["link"],
-                                channel["title"], channel["url"], channel_subscribers,
-                                channel_views, channel["joinedDate"],
-                                (str(round(views / channel_subscribers * 100, 2)) + "%")))
-                    counter = counter + 1
-                    if counter == request_limit:
-                        break
-                if counter == request_limit:
-                    break
-                has_next_page = videos_search.next()
-            self._model.setData(result)
+        engine = YoutubeGrepEngine(self._model, request_limit)
+        if engine.search(self._request_text):
             self._table_view.resizeColumnsToContents()
-        except Exception as e:
+        else:
             dialog = QMessageBox()
             dialog.setWindowTitle(app_name)
             dialog.setText("Error in the searching process")
             dialog.setIcon(QMessageBox.Critical)
-            dialog.setDetailedText(traceback.format_exc())
+            dialog.setDetailedText(engine.error)
             dialog.exec()
         
         QApplication.restoreOverrideCursor()
